requirements/base.txt
django==1.8.11
pillow==3.1.1
django-stronghold==0.2.7
django-crispy-forms==1.6.0

django-authtools==1.5.0
# install authtools from github until they release the migration fix
#git+https://github.com/fusionbox/django-authtools.git@7f1a3e63f901ebe5f6ab9951bb51e1a62e53990f

django-tables2==1.1.2
django-braces==1.8.1
django-sekizai==0.9.0
fabric==1.10.2
logan==0.7.1
django-bootstrap-form==3.2
croniter==0.3.12
GitPython==1.0.2
gevent-socketio==0.3.6
django-graphos==0.1.1
django-activeurl==0.1.9
requests==2.9.1
virtualenv==15.0.0
ansiconv==1.0.0
channels==0.9.5
autobahn==0.12.1
twisted==16.4.0
six==1.10.0
pycrypto==2.6.1
